+++ src/Authority/Port.php
@@ -3,6 +3,9 @@ declare(strict_types = 1);
 
 namespace Innmind\Url\Authority;
 
+use Uri\WhatWg\Url as Concrete;
+use Uri\Rfc3986\Uri;
+
 /**
  * @psalm-immutable
  */
@@ -32,6 +35,21 @@ final class Port
     public static function none(): self
     {
         return new self(null);
+    }
+
+    /**
+     * @internal
+     * @psalm-pure
+     */
+    public static function parsed(Uri|Concrete $parsed): self
+    {
+        /** @psalm-suppress ImpureMethodCall */
+        $port = $parsed->getPort();
+
+        return match ($port) {
+            null => self::none(),
+            default => new self($port),
+        };
     }
 
     #[\NoDiscard]