+++ src/Path.php
@@ -4,6 +4,7 @@ declare(strict_types = 1);
 namespace Innmind\Url;
 
 use Uri\WhatWg\Url as Concrete;
+use Uri\Rfc3986\Uri;
 
 /**
  * @psalm-immutable
@@ -25,15 +26,16 @@ abstract class Path
         }
 
         try {
-            /** @psalm-suppress ImpureMethodCall */
-            $url = new Concrete('http://a.org/');
-            /** @psalm-suppress ImpureMethodCall */
-            $url = $url->withPath($value);
-
-            return $value[0] === '/' ? new AbsolutePath($value) : new RelativePath($value);
+            $path = Url::of($value)->path();
         } catch (\Exception) {
             throw new \DomainException($value);
         }
+
+        if ($path instanceof RelativePath) {
+            return new RelativePath($value);
+        }
+
+        return new AbsolutePath($value);
     }
 
     /**
@@ -43,6 +45,45 @@ abstract class Path
     final public static function none(): self
     {
         return new AbsolutePath('');
+    }
+
+    /**
+     * @internal
+     * @psalm-pure
+     */
+    final public static function parsed(Uri|Concrete $parsed): self
+    {
+        if ($parsed instanceof Uri) {
+            /** @psalm-suppress ImpureMethodCall */
+            $path = $parsed->getRawPath();
+        } else {
+            /** @psalm-suppress ImpureMethodCall */
+            $path = $parsed->getPath();
+        }
+
+        if ($path === '') {
+            return self::none();
+        }
+
+        if ($path[0] === '/') {
+            return new AbsolutePath($path);
+        }
+
+        return new RelativePath($path);
+    }
+
+    /**
+     * @internal
+     * @psalm-pure
+     */
+    final public static function initiallyRelative(self $self): self
+    {
+        $path = \substr($self->toString(), 1);
+
+        return match ($path) {
+            '' => self::none(),
+            default => new RelativePath($path),
+        };
     }
 
     #[\NoDiscard]