+++ src/Query.php
@@ -4,6 +4,7 @@ declare(strict_types = 1);
 namespace Innmind\Url;
 
 use Uri\WhatWg\Url as Concrete;
+use Uri\Rfc3986\Uri;
 
 /**
  * @psalm-immutable
@@ -21,13 +22,7 @@ final class Query
     public static function of(string $value): self
     {
         try {
-            /** @psalm-suppress ImpureMethodCall */
-            $url = new Concrete('http://a.org');
-            /** @psalm-suppress ImpureMethodCall */
-            $url = $url->withQuery($value);
-
-            /** @psalm-suppress ImpureMethodCall */
-            return new self((string) $url->getQuery());
+            return Url::of('http://a.org/?'.$value)->query();
         } catch (\Exception) {
             throw new \DomainException($value);
         }
@@ -40,6 +35,21 @@ final class Query
     public static function none(): self
     {
         return new self('');
+    }
+
+    /**
+     * @internal
+     * @psalm-pure
+     */
+    public static function parsed(Uri|Concrete $parsed): self
+    {
+        /** @psalm-suppress ImpureMethodCall */
+        $query = $parsed->getQuery();
+
+        return match ($query) {
+            null => self::none(),
+            default => new self($query),
+        };
     }
 
     #[\NoDiscard]